broke down original method into manageable parts
