+++ src/__init__.py
@@ -1,0 +1,3 @@
+"""
+Force Board Metrics - Balance Board Application
+"""
