+++ src/style.qss
@@ -1,0 +1,14 @@
+QMainWindow {
+    background-color: #2b2b2b;
+}
+
+QLabel {
+    color: #ffffff;
+    font-size: 14px;
+    padding: 10px;
+}
+
+QLabel#weight_label {
+    font-size: 24px;
+    font-weight: bold;
+}
